Add plot_training_history() function. Add 2nd test scenario
